remove_the_circular_reference_annotation_of_edit_bookmark_modal
