+++ src/contexts/global_context.tsx
@@ -7,9 +7,7 @@ import { LAYOUT_BREAKPOINT } from '@/constants/display';
 import { LayoutAssertion } from '@/interfaces/layout_assertion';
 import ConfirmModal from '@/components/confirm_modal/confirm_modal';
 import AddAssetModal from '@/components/add_asset_modal/add_asset_modal';
-// eslint-disable-next-line import/no-cycle
 import CameraScanner from '@/components/camera_scanner/camera_scanner';
-// eslint-disable-next-line import/no-cycle
 import PreviewInvoiceModal from '@/components/preview_invoice_modal/preview_invoice_modal';
 import {
   IPreviewInvoiceModal,
@@ -18,9 +16,7 @@ import {
 import EmbedCodeModal from '@/components/embed_code_modal/embed_code_modal';
 import Toast from '@/components/toast/toast';
 import { ToastPosition, ToastType } from '@/interfaces/toastify';
-// eslint-disable-next-line import/no-cycle
 import CreateCompanyModal from '@/components/create_company_modal/create_company_modal';
-// eslint-disable-next-line import/no-cycle
 import CompanyInvitationModal from '@/components/company_invitation_modal/company_invitation_modal';
 import Link from 'next/link';
 import { ISUNFA_ROUTE } from '@/constants/url';
@@ -28,25 +24,18 @@ import { useRouter } from 'next/router';
 import LoadingModal from '@/components/loading_modal/loading_modal';
 import { IConfirmModal, dummyConfirmModalData } from '@/interfaces/confirm_modal';
 import FilterOptionsModal from '@/components/filter_options_modal/filter_options_modal';
-// eslint-disable-next-line import/no-cycle
 import AddProjectModal from '@/components/add_project_modal/add_project_modal';
 import { useUserCtx } from '@/contexts/user_context';
 import { useNotificationCtx } from '@/contexts/notification_context';
 import { ProjectStage } from '@/constants/project';
-// eslint-disable-next-line import/no-cycle
 import EditBookmarkModal from '@/components/edit_bookmark_modal/edit_bookmark_modal';
-// eslint-disable-next-line import/no-cycle
 import ProfileUploadModal from '@/components/profile_upload_modal/profile_upload_modal';
 import SalaryBookConfirmModal from '@/components/salary_book_confirm_modal/salary_book_confirm_modal';
 import { ToastId } from '@/constants/toast_id';
 import { useTranslation } from 'next-i18next';
-// eslint-disable-next-line import/no-cycle
 import AddAccountTitleModal from '@/components/add_account_title_modal/add_account_title_modal';
-// eslint-disable-next-line import/no-cycle
 import EditAccountTitleModal from '@/components/edit_account_title_modal/edit_account_title_modal';
-// eslint-disable-next-line import/no-cycle
 import TeamSettingModal from '@/components/team_setting_modal/team_setting_modal';
-// eslint-disable-next-line import/no-cycle
 import TransferCompanyModal from '@/components/transfer_company_modal/transfer_company_modal';
 import { UploadType } from '@/constants/file';
 import LoginConfirmModal from '@/components/login_confirm_modal/login_confirm_modal';
@@ -60,16 +49,6 @@ interface IGlobalContext {
   isPasskeySupportModalVisible: boolean;
   passKeySupportModalVisibilityHandler: () => void;
 
-  isAddBookmarkModalVisible: boolean;
-  addBookmarkModalVisibilityHandler: () => void;
-
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
-  // isMessageModalVisible: boolean;
-  // messageModalVisibilityHandler: () => void;
-  // messageModalData: IMessageModal;
-  // messageModalDataHandler: (data: IMessageModal) => void;
-  // isConfirmModalVisible: boolean;
-  // confirmModalVisibilityHandler: () => void;
   confirmModalData: IConfirmModal;
   confirmModalDataHandler: (data: IConfirmModal) => void;
 
@@ -114,10 +93,6 @@ interface IGlobalContext {
   editAccountTitleModalVisibilityHandler: () => void;
   editAccountTitleDataHandler: (id: number) => void;
 
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
-  // toastHandler: (props: IToastify) => void;
-  // eliminateToast: (id?: string) => void;
-
   filterOptionsForHistory: IFilterOptions;
   filterOptionsForPending: IFilterOptions;
   filterOptionsForContract: IFilterOptions;
@@ -154,7 +129,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
   const { reportGeneratedStatus, reportPendingStatus, reportGeneratedStatusHandler } =
     useNotificationCtx();
 
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部沒問題時，再將這個註解拿掉
   const {
     toastHandler,
     eliminateToast,
@@ -163,17 +137,13 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
     messageModalData,
     isConfirmModalVisible,
     confirmModalVisibilityHandler,
+    isAddBookmarkModalVisible,
+    addBookmarkModalVisibilityHandler,
   } = useModalContext();
 
   const windowSize = useWindowSize();
   const [isPasskeySupportModalVisible, setIsPasskeySupportModalVisible] = useState(false);
 
-  const [isAddBookmarkModalVisible, setIsAddBookmarkModalVisible] = useState(false);
-
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
-  // const [isMessageModalVisible, setIsMessageModalVisible] = useState(false);
-  // const [messageModalData, setMessageModalData] = useState<IMessageModal>(dummyMessageModalData);
-  // const [isConfirmModalVisible, setIsConfirmModalVisible] = useState(false);
   const [confirmModalData, setConfirmModalData] = useState<IConfirmModal>(dummyConfirmModalData);
 
   const [isAddAssetModalVisible, setIsAddAssetModalVisible] = useState(false);
@@ -242,24 +212,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
   const passKeySupportModalVisibilityHandler = () => {
     setIsPasskeySupportModalVisible(!isPasskeySupportModalVisible);
   };
-
-  const addBookmarkModalVisibilityHandler = () => {
-    setIsAddBookmarkModalVisible(!isAddBookmarkModalVisible);
-  };
-
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
-  // const messageModalVisibilityHandler = () => {
-  //   setIsMessageModalVisible(!isMessageModalVisible);
-  // };
-
-  // const messageModalDataHandler = (data: IMessageModal) => {
-  //   setMessageModalData(data);
-  // };
-
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
-  // const confirmModalVisibilityHandler = () => {
-  //   setIsConfirmModalVisible(!isConfirmModalVisible);
-  // };
 
   const confirmModalDataHandler = (data: IConfirmModal) => {
     setConfirmModalData(data);
@@ -382,118 +334,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
     setIsTOSNPrivacyPolicyConfirmModalVisible(visibility);
   };
 
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
-  // Info: (20240509 - Julian) toast handler
-  // const toastHandler = useCallback((props: IToastify) => {
-  //   const {
-  //     // eslint-disable-next-line react/prop-types
-  //     id,
-  //     // eslint-disable-next-line react/prop-types
-  //     type,
-  //     // eslint-disable-next-line react/prop-types
-  //     content,
-  //     // eslint-disable-next-line react/prop-types
-  //     closeable,
-  //     // eslint-disable-next-line react/prop-types
-  //     autoClose: isAutoClose,
-  //     // eslint-disable-next-line react/prop-types
-  //     position: toastPosition,
-  //     // eslint-disable-next-line react/prop-types
-  //     onClose = () => {},
-  //     // eslint-disable-next-line react/prop-types
-  //     onOpen = () => {},
-  //   } = props;
-
-  //   const bodyStyle =
-  //     'before:absolute before:h-100vh before:w-5px before:top-0 before:left-0 md:w-400px w-100vw md:scale-100 scale-75 text-sm font-barlow pointer-events-auto';
-
-  //   const toastId = id;
-  //   const position = toastPosition ?? ToastPosition.TOP_CENTER; // Info:(20240513 - Julian) default position 'top-center'
-
-  //   // Info:(20240513 - Julian) 如果 closeable 為 false，則 autoClose、closeOnClick、draggable 都會被設為 false
-  //   const autoClose = closeable ? (isAutoClose ?? 5000) : false; // Info:(20240513 - Julian) default autoClose 5000ms
-
-  //   const closeOnClick = closeable; // Info:(20240513 - Julian) default closeOnClick true
-  //   const draggable = closeable; // Info:(20240513 - Julian) default draggable true
-  //   const closeButton = closeable
-  //     ? () => (
-  //         <div className="h-20px w-20px">
-  //           <RxCross2 size={16} className="text-button-text-secondary" />
-  //         </div>
-  //       )
-  //     : false;
-
-  //   switch (type) {
-  //     case ToastType.SUCCESS:
-  //       toastify.success(content, {
-  //         icon: <Image src="/icons/success.svg" alt="success" width={24} height={24} />,
-  //         className: `${bodyStyle} before:bg-alert-surface-surface-success`,
-  //         toastId,
-  //         position,
-  //         autoClose,
-  //         closeOnClick,
-  //         draggable,
-  //         closeButton,
-  //         onClose,
-  //         onOpen,
-  //       });
-  //       break;
-  //     case ToastType.ERROR:
-  //       toastify.error(content, {
-  //         icon: <Image src="/icons/error.svg" alt="error" width={24} height={24} />,
-  //         className: `${bodyStyle} before:bg-alert-surface-surface-error`,
-  //         toastId,
-  //         position,
-  //         autoClose,
-  //         closeOnClick,
-  //         draggable,
-  //         closeButton,
-  //         onClose,
-  //         onOpen,
-  //       });
-  //       break;
-  //     case ToastType.WARNING:
-  //       toastify.warning(content, {
-  //         icon: <Image src="/icons/warning.svg" alt="warning" width={24} height={24} />,
-  //         className: `${bodyStyle} before:bg-alert-surface-surface-warning`,
-  //         toastId,
-  //         position,
-  //         autoClose,
-  //         closeOnClick,
-  //         draggable,
-  //         closeButton,
-  //         onClose,
-  //         onOpen,
-  //       });
-  //       break;
-  //     case ToastType.INFO:
-  //       toastify.info(content, {
-  //         icon: <Image src="/icons/info.svg" alt="info" width={24} height={24} />,
-  //         className: `${bodyStyle} before:bg-alert-surface-surface-info`,
-  //         toastId,
-  //         position,
-  //         autoClose,
-  //         closeOnClick,
-  //         draggable,
-  //         closeButton,
-  //         onClose,
-  //         onOpen,
-  //       });
-  //       break;
-  //     default:
-  //       toastify(content);
-  //       break;
-  //   }
-  // }, []);
-
-  // const eliminateToast = (id?: string) => {
-  //   if (id) {
-  //     toastify.dismiss(id);
-  //   } else {
-  //     toastify.dismiss(); // Info:(20240513 - Julian) dismiss all toasts
-  //   }
-  // };
-
   useEffect(() => {
     if (!signedIn) return;
 
@@ -605,7 +445,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
 
   // Info: (20240830 - Anna) 為了拿掉react/jsx-no-constructed-context-values註解，所以使用useMemo hook
 
-  // Todo: (20240909 - Anna) 代碼正在搬動到src/contexts/modal_context.tsx當中，確定全部搬完時，再將這段代碼刪除
   const value = useMemo(
     () => ({
       width,
@@ -613,14 +452,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
       layoutAssertion,
       isPasskeySupportModalVisible,
       passKeySupportModalVisibilityHandler,
-      isAddBookmarkModalVisible,
-      addBookmarkModalVisibilityHandler,
-      // isMessageModalVisible,
-      // messageModalVisibilityHandler,
-      // messageModalData,
-      // messageModalDataHandler,
-      // isConfirmModalVisible,
-      // confirmModalVisibilityHandler,
       confirmModalData,
       confirmModalDataHandler,
       isAddAssetModalVisible,
@@ -646,8 +477,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
       isEditAccountTitleModalVisible,
       editAccountTitleModalVisibilityHandler,
       editAccountTitleDataHandler,
-      // toastHandler,
-      // eliminateToast,
 
       filterOptionsForHistory,
       filterOptionsForPending,
@@ -680,14 +509,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
       layoutAssertion,
       isPasskeySupportModalVisible,
       passKeySupportModalVisibilityHandler,
-      isAddBookmarkModalVisible,
-      addBookmarkModalVisibilityHandler,
-      // isMessageModalVisible,
-      // messageModalVisibilityHandler,
-      // messageModalData,
-      // messageModalDataHandler,
-      // isConfirmModalVisible,
-      // confirmModalVisibilityHandler,
       confirmModalData,
       confirmModalDataHandler,
       isAddAssetModalVisible,
@@ -713,8 +534,6 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
       isEditAccountTitleModalVisible,
       editAccountTitleModalVisibilityHandler,
       editAccountTitleDataHandler,
-      // toastHandler,
-      // eliminateToast,
 
       filterOptionsForHistory,
       filterOptionsForPending,
@@ -749,7 +568,7 @@ export const GlobalProvider = ({ children }: IGlobalProvider) => {
         isModalVisible={isPasskeySupportModalVisible}
         modalVisibilityHandler={passKeySupportModalVisibilityHandler}
       />
-      {/* Todo: (20240829 - Anna) 修改循環引用問題 */}
+
       <EditBookmarkModal
         isModalVisible={isAddBookmarkModalVisible}
         modalVisibilityHandler={addBookmarkModalVisibilityHandler}
